+++ src/components/alert/index.ts
@@ -1,25 +1,2 @@
-import { Component, Input } from '@angular/core';
+export * from './alert.component';
 
-@Component({
-  selector: 'rio-alert',
-  template: `
-    <div
-      [id]="qaid"
-      class="p2 bold"
-      [attr.data-testid]="testid"
-      [ngClass]="{
-        'bg-blue': status === 'info',
-        'bg-yellow': status === 'warning',
-        'bg-green': status === 'success',
-        'bg-red': status === 'error',
-        'white': status === 'info' || status === 'error'
-      }">
-      <ng-content></ng-content>
-    </div>
-  `
-})
-export class RioAlert {
-  @Input() status = 'info';
-  @Input() qaid: string;
-  @Input() testid: string;
-};
